move selfplay into its own folder
