+++ cc/selfplay/self_play_thread.h
